EPIC-1156 MEM-440 Prevent deletion of published content.
The main new rule is to not let the user delete published content. Existing rules include prevention of folders with content and user permissions.
There are 5 ways to delete content. 
# Info panel. Select a document and open the info panel. Try deleting a published and unpublished file.
# Row item delete. Open the ellipse menu for a published document. Click the delete icon. Repeat for a unpublished document.
# Repeat #2 on a published and unpublished folder.
# Select various files and folders and use the batch menu delete (top right side above document list).

Scenarios to test:
# Only published folders
# Only published files
# Only unpublished folders
# Only unpublished files
# Mix published folders and unpublished folders
# Mix published files and unpublished files
# Mix published and unpublished folders and files.
# Only folders with content
# Repeat above mixing in folders with content

This dialog also handles the case a user is not authorized to delete a file or files and the case a user can not manage folders.

Note that the system has a defect with multiple folder deletes. After a multiple delete operation the user must refresh the page to get the results.